container_image/requirements.txt
fs
redis
python-dotenv
